server/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.92
docutils==0.21.2
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.65.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
keras==3.4.1
keyring==25.2.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.3.0
mtcnn==0.1.1
namex==0.0.8
nh3==0.2.18
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
pkginfo==1.10.0
protobuf==4.25.3
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==2.0.0
rich==13.7.1
six==1.16.0
soupsieve==2.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
tqdm==4.66.4
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
